+++ include/autd3/driver/common/fpga/defined.hpp
@@ -3,7 +3,7 @@
 // Created Date: 10/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 09/11/2022
+// Last Modified: 15/11/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -25,25 +25,11 @@ namespace autd3::driver {
 
 constexpr size_t FPGA_CLK_FREQ = 163840000;
 
-constexpr uint16_t MAX_CYCLE = 8191;
-
-constexpr uint32_t MOD_SAMPLING_FREQ_DIV_MIN = 580;
-constexpr size_t MOD_BUF_SIZE_MAX = 65536;
-
 #ifdef AUTD3_USE_METER
 constexpr double POINT_STM_FIXED_NUM_UNIT = 0.025e-3;
 #else
 constexpr double POINT_STM_FIXED_NUM_UNIT = 0.025;
 #endif
-
-constexpr uint32_t POINT_STM_SAMPLING_FREQ_DIV_MIN = 806;
-constexpr uint32_t GAIN_STM_SAMPLING_FREQ_DIV_MIN = 138;
-constexpr uint32_t GAIN_STM_LEGACY_SAMPLING_FREQ_DIV_MIN = 76;
-constexpr size_t POINT_STM_BUF_SIZE_MAX = 65536;
-constexpr size_t GAIN_STM_BUF_SIZE_MAX = 1024;
-constexpr size_t GAIN_STM_LEGACY_BUF_SIZE_MAX = 2048;
-
-constexpr uint16_t SILENCER_CYCLE_MIN = 522;
 
 #ifdef _MSC_VER
 #pragma warning(push)