+++ include/autd3/soft_stm.hpp
@@ -3,7 +3,7 @@
 // Created Date: 07/09/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 07/11/2022
+// Last Modified: 15/11/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -80,16 +80,16 @@ class SoftwareSTM {
       if (!_run) throw std::runtime_error("STM has been already finished.");
       _run = false;
       if (_th.joinable()) _th.join();
-      _cnt.check_trials = _trials;
+      _cnt.set_ack_check_timeout(_timeout);
     }
 
    private:
     SoftwareSTMThreadHandle(Controller& cnt, const std::vector<std::shared_ptr<core::Gain>>& bodies, const uint64_t period,
                             const TimerStrategy strategy)
-        : _cnt(cnt), _trials(_cnt.check_trials) {
+        : _cnt(cnt), _timeout(_cnt.get_ack_check_timeout()) {
       _run = true;
       const auto interval = std::chrono::nanoseconds(period);
-      _cnt.check_trials = 0;
+      _cnt.set_ack_check_timeout(std::chrono::high_resolution_clock::duration::zero());
       if (strategy.contains(TimerStrategy::BUSY_WAIT))
         _th = std::thread([this, interval, bodies]() {
           size_t i = 0;
@@ -120,7 +120,7 @@ class SoftwareSTM {
     bool _run;
     std::thread _th;
     Controller& _cnt;
-    size_t _trials;
+    std::chrono::high_resolution_clock::duration _timeout;
   };
 
   SoftwareSTM() noexcept : timer_strategy(TimerStrategy::NONE), _sample_period_ns(0) {}