+++ python/pyautd3/autd.py
@@ -4,7 +4,7 @@ Project: pyautd3
 Created Date: 24/05/2021
 Author: Shun Suzuki
 -----
-Last Modified: 08/11/2022
+Last Modified: 15/11/2022
 Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 -----
 Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -26,6 +26,13 @@ NUM_TRANS_Y = 14
 TRANS_SPACING = 10.16
 DEVICE_WIDTH = 192.0
 DEVICE_HEIGHT = 151.4
+
+DRIVER_LATEST = 0x00
+DRIVER_V2_2 = 0x82
+DRIVER_V2_3 = 0x83
+DRIVER_V2_4 = 0x84
+DRIVER_V2_5 = 0x85
+DRIVER_V2_6 = 0x86
 
 
 class SpecialData:
@@ -219,9 +226,9 @@ class Geometry:
 
 
 class Controller:
-    def __init__(self):
+    def __init__(self, driver_version: int = DRIVER_LATEST):
         self.p_cnt = c_void_p()
-        Base().dll.AUTDCreateController(byref(self.p_cnt))
+        Base().dll.AUTDCreateController(byref(self.p_cnt), driver_version)
         self.__disposed = False
 
     def __del__(self):
@@ -288,19 +295,35 @@ class Controller:
         return Base().dll.AUTDSetForceFan(self.p_cnt, value)
 
     @ property
-    def check_trials(self):
-        return Base().dll.AUTDGetCheckTrials(self.p_cnt)
+    def ack_check_timeout_ms(self):
+        return Base().dll.AUTDGetAckCheckTimeout(self.p_cnt) / 1000 / 1000
 
-    @ check_trials.setter
-    def check_trials(self, value: int):
-        return Base().dll.AUTDSetCheckTrials(self.p_cnt, value)
+    @ ack_check_timeout_ms.setter
+    def ack_check_timeout_ms(self, value: int):
+        return Base().dll.AUTDSetAckCheckTimeout(self.p_cnt, value * 1000 * 1000)
 
     @ property
-    def send_interval(self):
+    def ack_check_timeout_ns(self):
+        return Base().dll.AUTDGetAckCheckTimeout(self.p_cnt)
+
+    @ ack_check_timeout_ns.setter
+    def ack_check_timeout_ns(self, value: int):
+        return Base().dll.AUTDSetAckCheckTimeout(self.p_cnt, value)
+
+    @ property
+    def send_interval_ms(self):
+        return Base().dll.AUTDGetSendInterval(self.p_cnt) / 1000 / 1000
+
+    @ send_interval_ms.setter
+    def send_interval_ms(self, value: int):
+        return Base().dll.AUTDSetSendInterval(self.p_cnt, value * 1000 * 1000)
+
+    @ property
+    def send_interval_ns(self):
         return Base().dll.AUTDGetSendInterval(self.p_cnt)
 
-    @ send_interval.setter
-    def send_interval(self, value: int):
+    @ send_interval_ns.setter
+    def send_interval_ns(self, value: int):
         return Base().dll.AUTDSetSendInterval(self.p_cnt, value)
 
     @ property