+++ src/link/soem/link_soem_impl.hpp
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 05/11/2022
+// Last Modified: 15/11/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -18,8 +18,7 @@
 #include <utility>
 
 #include "autd3/core/link.hpp"
-#include "autd3/driver/cpu/datagram.hpp"
-#include "autd3/driver/cpu/header.hpp"
+#include "autd3/driver/common/cpu/datagram.hpp"
 #include "autd3/link/soem.hpp"
 #include "soem_handler.hpp"